Applying (rebased) HIVE-1943 to r7.

+++ metastore/src/java/org/apache/hadoop/hive/metastore/HiveMetaStore.java
@@ -531,7 +531,8 @@ public class HiveMetaStore extends ThriftHiveMetastore {
     }
 
     private void create_database_core(RawStore ms, final Database db)
-        throws AlreadyExistsException, InvalidObjectException, MetaException {
+        throws AlreadyExistsException, InvalidObjectException, MetaException,
+        IOException {
       if (!validateName(db.getName())) {
         throw new InvalidObjectException(db.getName() + " is not a valid database name");
       }
@@ -637,7 +638,8 @@ public class HiveMetaStore extends ThriftHiveMetastore {
 
     private void drop_database_core(RawStore ms,
         final String name, final boolean deleteData)
-        throws NoSuchObjectException, InvalidOperationException, MetaException {
+        throws NoSuchObjectException, InvalidOperationException, MetaException,
+        IOException {
       boolean success = false;
       Database db = null;
       try {
@@ -645,6 +647,12 @@ public class HiveMetaStore extends ThriftHiveMetastore {
         db = ms.getDatabase(name);
         if (!get_all_tables(db.getName()).isEmpty()) {
           throw new InvalidOperationException("Database " + db.getName() + " is not empty");
+        }
+        Path path = new Path(db.getLocationUri()).getParent();
+        if (!wh.isWritable(path)) {
+          throw new MetaException("Database not dropped since " +
+              path + " is not writable by " + 
+              hiveConf.getUser());
         }
         if (ms.dropDatabase(name)) {
           success = ms.commitTransaction();
@@ -967,7 +975,7 @@ public class HiveMetaStore extends ThriftHiveMetastore {
 
     private void drop_table_core(final RawStore ms, final String dbname,
         final String name, final boolean deleteData)
-        throws NoSuchObjectException, MetaException {
+        throws NoSuchObjectException, MetaException, IOException {
 
       boolean success = false;
       boolean isExternal = false;
@@ -1008,6 +1016,11 @@ public class HiveMetaStore extends ThriftHiveMetastore {
         isExternal = isExternal(tbl);
         if (tbl.getSd().getLocation() != null) {
           tblPath = new Path(tbl.getSd().getLocation());
+          if (!wh.isWritable(tblPath.getParent())) {
+            throw new MetaException("Table metadata not deleted since " +
+                tblPath.getParent() + " is not writable by " + 
+                hiveConf.getUser());
+          }
         }
 
         if (!ms.dropTable(dbname, name)) {
@@ -1348,7 +1361,7 @@ public class HiveMetaStore extends ThriftHiveMetastore {
 
     private boolean drop_partition_common(RawStore ms, String db_name, String tbl_name,
         List<String> part_vals, final boolean deleteData)
-    throws MetaException, NoSuchObjectException {
+    throws MetaException, NoSuchObjectException, IOException {
 
       boolean success = false;
       Path partPath = null;
@@ -1369,6 +1382,11 @@ public class HiveMetaStore extends ThriftHiveMetastore {
         isArchived = MetaStoreUtils.isArchived(part);
         if (isArchived) {
           archiveParentDir = MetaStoreUtils.getOriginalLocation(part);
+          if (!wh.isWritable(archiveParentDir.getParent())) {
+            throw new MetaException("Table partition not deleted since " +
+                archiveParentDir.getParent() + " is not writable by " + 
+                hiveConf.getUser());
+          }
         }
         if (part.getSd() == null || part.getSd().getLocation() == null) {
           throw new MetaException("Partition metadata is corrupted");
@@ -1378,6 +1396,11 @@ public class HiveMetaStore extends ThriftHiveMetastore {
         }
         success = ms.commitTransaction();
         partPath = new Path(part.getSd().getLocation());
+        if (!wh.isWritable(partPath.getParent())) {
+           throw new MetaException("Table partition not deleted since " +
+                partPath.getParent() + " is not writable by " + 
+                 hiveConf.getUser());
+        }
         tbl = get_table(db_name, tbl_name);
       } finally {
         if (!success) {
@@ -1950,7 +1973,7 @@ public class HiveMetaStore extends ThriftHiveMetastore {
     private boolean drop_partition_by_name_core(final RawStore ms,
         final String db_name, final String tbl_name, final String part_name,
         final boolean deleteData) throws NoSuchObjectException,
-        MetaException, TException {
+        MetaException, TException, IOException {
 
       List<String> partVals = null;
       try {
@@ -2219,7 +2242,7 @@ public class HiveMetaStore extends ThriftHiveMetastore {
     private boolean drop_index_by_name_core(final RawStore ms,
         final String dbName, final String tblName,
         final String indexName, final boolean deleteData) throws NoSuchObjectException,
-        MetaException, TException {
+        MetaException, TException, IOException {
 
       boolean success = false;
       Path tblPath = null;
@@ -2243,6 +2266,11 @@ public class HiveMetaStore extends ThriftHiveMetastore {
 
           if (tbl.getSd().getLocation() != null) {
             tblPath = new Path(tbl.getSd().getLocation());
+            if (!wh.isWritable(tblPath.getParent())) {
+              throw new MetaException("Index table metadata not deleted since " +
+                  tblPath.getParent() + " is not writable by " + 
+                  hiveConf.getUser());
+            }
           }
           if (!ms.dropTable(dbName, idxTblName)) {
             throw new MetaException("Unable to drop underlying data table "
